server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.1
yt-dlp==2023.12.30
websockets==12.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-socketio==5.11.1
spotipy==2.23.0
certifi==2024.2.2
requests==2.31.0
pytube==15.0.0
aiohttp==3.9.3
# Add additional dependencies for YouTube downloads
beautifulsoup4==4.12.3
lxml==5.1.0 